created basic whiteboard application w/keybinds + blocks
I followed several tutorials online to do this. Since I was typing while following along, it took a bit longer than an hour for my first commit, but I wanted to finish all the blocks. I will see if I can link the tutorials on the readme. I also looked up a list of items that should be cutomizable for a whiteboard application. excited to add functionality.